requirements.txt
streamlit
Pillow
pytesseract
pdfplumber
transformers
sentence-transformers
torch
faiss-cpu
